Move group components into new group module and updates to tests
