+++ docs/source/swan/components/group.rst
@@ -1,0 +1,14 @@
+================
+Group
+================
+
+SWAN group commands
+
+.. autosummary::
+   :nosignatures:
+   :toctree: _generated/
+
+   rompy.swan.components.group.BaseGroupComponent
+   rompy.swan.components.group.STARTUP
+   rompy.swan.components.group.PHYSICS
+   rompy.swan.components.group.OUTPUT
