+++ rompy/swan/components/output.py
@@ -1628,276 +1628,3 @@ class TEST(BaseComponent):
         if self.fname_s2d is not None:
             repr += f"S2D fname='{self.fname_s2d}' "
         return repr.rstrip()
-
-
-# =====================================================================================
-# Output group component
-# =====================================================================================
-FRAME_TYPE = Annotated[FRAME, Field(description="Frame locations component")]
-GROUP_TYPE = Annotated[GROUP, Field(description="Group locations component")]
-RAY_TYPE = Annotated[RAY, Field(description="Ray locations component")]
-ISOLINE_TYPE = Annotated[ISOLINE, Field(description="Isoline locations component")]
-OUTOPT_TYPE = Annotated[OUTPUT_OPTIONS, Field(description="Output options component")]
-BLOCK_TYPE = Annotated[BLOCK, Field(description="Block write component")]
-TABLE_TYPE = Annotated[TABLE, Field(description="Table write component")]
-SPECOUT_TYPE = Annotated[SPECOUT, Field(description="Spectra write component")]
-NESTOUT_TYPE = Annotated[NESTOUT, Field(description="Nest write component")]
-TEST_TYPE = Annotated[TEST, Field(description="Intermediate write component")]
-CURVE_TYPES = Annotated[
-    Union[CURVE, CURVES],
-    Field(description="Curve locations component", discriminator="model_type"),
-]
-POINTS_TYPES = Annotated[
-    Union[POINTS, POINTS_FILE],
-    Field(description="Points locations component", discriminator="model_type"),
-]
-NGRID_TYPES = Annotated[
-    Union[NGRID, NGRID_UNSTRUCTURED],
-    Field(description="Ngrid locations component", discriminator="model_type")
-]
-QUANTITY_TYPES = Annotated[
-    Union[QUANTITY, QUANTITIES],
-    Field(description="Quantity component", discriminator="model_type")
-]
-
-
-class OUTPUT(BaseComponent):
-    """Output group component.
-
-    .. code-block:: text
-
-        FRAME 'sname' ...
-        GROUP 'sname' ...
-        CURVE 'sname' ...
-        RAY 'rname' ...
-        ISOLINE 'sname' 'rname' ...
-        POINTS 'sname ...
-        NGRID 'sname' ...
-        QUANTITY ...
-        OUTPUT OPTIONS ...
-        BLOCK 'sname' ...
-        TABLE 'sname' ...
-        SPECOUT 'sname' ...
-        NESTOUT 'sname ...
-
-    This group component is used to define multiple types of output locations and
-    write components in a single model. Only fields that are explicitly prescribed are
-    rendered by this group component.
-
-    Note
-    ----
-    The components prescribed are validated according to some constraints as defined
-    in the SWAN manual:
-
-    - The name `'sname'` of each Locations component must be unique.
-    - The Locations `'sname'` assigned to each write component must be defined.
-    - The BLOCK component must be associated with either a `FRAME` or `GROUP`.
-    - The ISOLINE write component must be associated with a `RAY` component.
-    - The NGRID and NESTOUT components must be defined together.
-
-    Examples
-    --------
-
-    .. ipython:: python
-        :okwarning:
-
-        from rompy.swan.components.output import POINTS, BLOCK, QUANTITY, TABLE, OUTPUT
-        points = POINTS(sname="outpts", xp=[172.3, 172.4], yp=[-39, -39])
-        quantity = QUANTITY(output=["depth", "hsign", "tps", "dir", "tm01"], excv=-9)
-        times = dict(
-            tbeg=dict(time="2012-01-01T00:00:00", tfmt=1),
-            delt=dict(delt="PT30M", dfmt="min"),
-        )
-        block = BLOCK(
-            model_type="block",
-            sname="COMPGRID",
-            fname="./swangrid.nc",
-            output=["depth", "hsign", "tps", "dir"],
-            time=times,
-        )
-        table = TABLE(
-            sname="outpts",
-            format="noheader",
-            fname="./swantable.nc",
-            output=["hsign", "hswell", "dir", "tps", "tm01", "watlev", "qp"],
-            time=times,
-        )
-        out = OUTPUT(
-            points=points,
-            quantity=quantity,
-            block=block,
-            table=table,
-        )
-        print(out.render())
-
-    """
-
-    model_type: Literal["output", "OUTPUT"] = Field(
-        default="output", description="Model type discriminator"
-    )
-    frame: Optional[FRAME_TYPE] = Field(default=None)
-    group: Optional[GROUP_TYPE] = Field(default=None)
-    curve: Optional[CURVE_TYPES] = Field(default=None)
-    ray: Optional[RAY_TYPE] = Field(default=None)
-    isoline: Optional[ISOLINE_TYPE] = Field(default=None)
-    points: Optional[POINTS_TYPES] = Field(default=None)
-    ngrid: Optional[NGRID_TYPES] = Field(default=None)
-    quantity: Optional[QUANTITY_TYPES] = Field(default=None)
-    output_options: Optional[OUTOPT_TYPE] = Field(default=None)
-    block: Optional[BLOCK_TYPE] = Field(default=None)
-    table: Optional[TABLE_TYPE] = Field(default=None)
-    specout: Optional[SPECOUT_TYPE] = Field(default=None)
-    nestout: Optional[NESTOUT_TYPE] = Field(default=None)
-    test: Optional[TEST_TYPE] = Field(default=None)
-    _location_fields: list = ["frame", "group", "curve", "isoline", "points", "ngrid"]
-    _write_fields: list = ["block", "table", "specout", "nestout"]
-
-    @model_validator(mode="after")
-    def write_locations_exists(self) -> "OUTPUT":
-        """Ensure the location component requested by a write component exists."""
-        for write in self.write_set:
-            sname = getattr(self, write).sname
-            if sname in SPECIAL_NAMES:
-                return self
-            try:
-                self._filter_location(sname)
-            except ValueError as err:
-                raise ValueError(
-                    f"Write component '{write}' specified with sname='{sname}' but no "
-                    f"location component with sname='{sname}' has been defined"
-                ) from err
-        return self
-
-    @model_validator(mode="after")
-    def locations_sname_unique(self) -> "OUTPUT":
-        """Ensure same `sname` isn't used in more than one set of output locations."""
-        duplicates = {x for x in self.snames if self.snames.count(x) > 1}
-        if duplicates:            
-            raise ValueError(
-                "The following snames are used to define more than one set of output "
-                f"components: {duplicates}, please ensure each location component has "
-                "a unique `sname`"
-            )
-        return self
-
-    @model_validator(mode="after")
-    def block_with_frame_or_group(self) -> "OUTPUT":
-        """Ensure Block is only defined for FRAME or GROUP locations."""
-        if self.block is not None:
-            sname = self.block.sname
-            if sname in ["BOTTGRID", "COMPGRID"]:
-                return self
-            location = self._filter_location(sname)
-            component = location.model_type.upper().split("_")[0]
-            if component not in ["FRAME", "GROUP"]:
-                raise ValueError(
-                    f"Block sname='{sname}' specified with {component} Location "
-                    "component but only only FRAME or GROUP components are supported"
-                )
-        return self
-
-    @model_validator(mode="after")
-    def isoline_ray_defined(self) -> "OUTPUT":
-        """Ensure the isoline ray has been defined."""
-        if self.isoline is not None:
-            if self.ray is None:
-                raise ValueError(
-                    f"Isoline {self.isoline} requires RAY rname='{self.isoline.rname}'"
-                    " but no RAY component has been defined"
-                )
-            elif self.ray.rname != self.isoline.rname:
-                raise ValueError(
-                    f"Isoline rname='{self.isoline.rname}' does not match "
-                    f"the ray rname='{self.ray.rname}'"
-                )
-        return self
-
-    @model_validator(mode="after")
-    def ngrid_and_nestout(self) -> "OUTPUT":
-        """Ensure NGRID and NESTOUT are specified together."""
-        if self.ngrid is not None and self.nestout is None:
-            raise ValueError(
-                "NGRID component specified but no NESTOUT component has been defined"
-            )
-        elif self.ngrid is None and self.nestout is not None:
-            raise ValueError(
-                "NESTOUT component specified but no NGRID component has been defined"
-            )
-        elif self.ngrid is not None and self.nestout is not None:
-            if self.ngrid.sname != self.nestout.sname:
-                raise ValueError(
-                    f"NGRID sname='{self.ngrid.sname}' does not match "
-                    f"the NESTOUT sname='{self.nestout.sname}'"
-                )
-        return self
-
-    @property
-    def locations_set(self):
-        """List of specified location fields."""
-        return [fld for fld in self.__fields_set__ if fld in self._location_fields]
-
-    @property
-    def write_set(self):
-        """List of specified write fields."""
-        return [fld for fld in self.__fields_set__ if fld in self._write_fields]
-
-    @property
-    def snames(self):
-        """List of snames from specified location components."""
-        snames = []
-        for field in self.locations_set:
-            sname = getattr(self, field).sname
-            if isinstance(sname, str):
-                sname = [sname]
-            snames.extend(sname)
-        return snames
-
-    def _filter_location(self, sname):
-        """Filter the location component defined with the specified sname."""
-        for field in self.locations_set:
-            obj = getattr(self, field)
-            obj_snames = obj.sname if isinstance(obj.sname, list) else [obj.sname]
-            for obj_sname in obj_snames:
-                if obj_sname == sname:
-                    return obj
-        raise ValueError(f"Location component with sname='{sname}' not found")
-
-    def cmd(self) -> str:
-        """Command file string for this component."""
-        repr = []
-        if self.frame is not None:
-            repr += [f"{self.frame.cmd()}"]
-        if self.group is not None:
-            repr += [f"{self.group.cmd()}"]
-        if self.curve is not None:
-            repr += [f"{self.curve.cmd()}"]
-        if self.ray is not None:
-            repr += [f"{self.ray.cmd()}"]
-        if self.isoline is not None:
-            repr += [f"{self.isoline.cmd()}"]
-        if self.points is not None:
-            repr += [f"{self.points.cmd()}"]
-        if self.ngrid is not None:
-            repr += [f"{self.ngrid.cmd()}"]
-        if self.quantity is not None:
-            repr += [f"{self.quantity.cmd()}"]
-        if self.output_options is not None:
-            repr += [f"{self.output_options.cmd()}"]
-        if self.block is not None:
-            repr += [f"{self.block.cmd()}"]
-        if self.table is not None:
-            repr += [f"{self.table.cmd()}"]
-        if self.specout is not None:
-            repr += [f"{self.specout.cmd()}"]
-        if self.nestout is not None:
-            repr += [f"{self.nestout.cmd()}"]
-        if self.test is not None:
-            repr += [f"{self.test.cmd()}"]
-        return repr
-
-    def render(self) -> str:
-        """Override base class to allow rendering list of components."""
-        cmds = []
-        for cmd in self.cmd():
-            cmds.append(super().render(cmd))
-        return "\n\n".join(cmds)
