+++ rompy/swan/components/physics.py
@@ -2884,10 +2884,10 @@ class OFF(BaseComponent):
         return f"OFF {self.physics.value.upper()}"
 
 
-OFF_TYPES = Annotated[list[OFF], Field(description="Deactivate physics")]
+OFF_TYPES = Annotated[list[OFF], Field(description="Deactivate physics command")]
 
 
-class DEACTIVATE(BaseComponent):
+class OFFS(BaseComponent):
     """Deactivate multiple physics commands.
 
     .. code-block:: text
@@ -2905,17 +2905,17 @@ class DEACTIVATE(BaseComponent):
     .. ipython:: python
         :okwarning:
 
-        from rompy.swan.components.physics import DEACTIVATE
+        from rompy.swan.components.physics import OFFS
         off1 = dict(physics="windgrowth")
         off2 = dict(physics="wcapping")
-        deactivate = DEACTIVATE(physics=[off1, off2])
-        for off in deactivate.render():
+        offs = OFFS(physics=[off1, off2])
+        for off in offs.render():
             print(off)
 
     """
 
-    model_type: Literal["deactivate", "DEACTIVATE"] = Field(
-        default="deactivate", description="Model type discriminator"
+    model_type: Literal["offs", "OFFS"] = Field(
+        default="offs", description="Model type discriminator"
     )
     physics: OFF_TYPES
 
@@ -2932,182 +2932,3 @@ class DEACTIVATE(BaseComponent):
         for cmd in self.cmd():
             cmds.append(super().render(cmd))
         return cmds
-
-
-# =====================================================================================
-# Physics group component
-# =====================================================================================
-GEN_TYPE = Annotated[
-    Union[GEN1, GEN2, GEN3],
-    Field(description="Wave generation component", discriminator="model_type"),
-]
-SSWELL_TYPE = Annotated[
-    Union[SSWELL_ROGERS, SSWELL_ARDHUIN, SSWELL_ZIEGER],
-    Field(description="Swell dissipation component", discriminator="model_type"),
-]
-NEGATINP_TYPE = Annotated[NEGATINP, Field(description="Negative wind input component")]
-WCAPPING_TYPE = Annotated[
-    Union[WCAPPING_KOMEN, WCAPPING_AB],
-    Field(description="Whitecapping component", discriminator="model_type"),
-]
-QUADRUPL_TYPE = Annotated[
-    QUADRUPL,
-    Field(description="Quadruplet interactions component"),
-]
-BREAKING_TYPE = Annotated[
-    Union[BREAKING_CONSTANT, BREAKING_BKD],
-    Field(description="Wave breaking component", discriminator="model_type"),
-]
-FRICTION_TYPE = Annotated[
-    Union[FRICTION_JONSWAP, FRICTION_COLLINS, FRICTION_MADSEN, FRICTION_RIPPLES],
-    Field(description="Bottom friction component", discriminator="model_type"),
-]
-TRIAD_TYPE = Annotated[
-    Union[TRIAD, TRIAD_DCTA, TRIAD_LTA, TRIAD_SPB],
-    Field(description="Triad interactions component", discriminator="model_type"),
-]
-VEGETATION_TYPE = Annotated[VEGETATION, Field(description="Vegetation component")]
-MUD_TYPE = Annotated[MUD, Field(description="Mud component")]
-SICE_TYPE = Annotated[
-    Union[SICE, SICE_R19, SICE_D15, SICE_M18, SICE_R21B],
-    Field(description="Sea ice component", discriminator="model_type"),
-]
-TURBULENCE_TYPE = Annotated[
-    TURBULENCE,
-    Field(description="Turbulent dissipation component"),
-]
-BRAGG_TYPE = Annotated[
-    Union[BRAGG, BRAGG_FT, BRAGG_FILE],
-    Field(description="Bragg scattering component", discriminator="model_type"),
-]
-LIMITER_TYPE = Annotated[LIMITER, Field(description="Limiter component")]
-OBSTACLE_TYPE = Annotated[OBSTACLES, Field(description="Obstacle group component")]
-SETUP_TYPE = Annotated[SETUP, Field(description="Setup component")]
-DIFFRACTION_TYPE = Annotated[DIFFRACTION, Field(description="Diffraction component")]
-SURFBEAT_TYPE = Annotated[SURFBEAT, Field(description="Surfbeat component")]
-SCAT_TYPE = Annotated[SCAT, Field(description="Scattering component")]
-DEACTIVATE_TYPE = Annotated[DEACTIVATE, Field(description="Deactivate group component")]
-
-
-class PHYSICS(BaseComponent):
-    """Physics group component.
-
-    The physics group component is a convenience to allow specifying several individual
-    components in a single command and check for consistency between them.
-
-    Exemples
-    --------
-
-    .. ipython:: python
-        :okwarning:
-
-        from rompy.swan.components.physics import PHYSICS
-        gen = {"model_type": "gen3", "source_terms": {"model_type": "komen"}}
-        phys = PHYSICS(gen=gen)
-        print(phys.render())
-        phys = PHYSICS(
-            gen=dict(model_type="gen3", source_terms={"model_type": "st6c1"}),
-            negatinp={"model_type": "negatinp", "rdcoef": 0.04},
-            sswell={"model_type": "zieger"},
-            breaking={"model_type": "constant", "alpha": 1.0, "gamma": 0.73},
-            friction={"model_type": "jonswap", "cfjon": 0.038},
-        )
-        print(phys.render())
-
-    """
-
-    model_type: Literal["physics", "PHYSICS"] = Field(
-        default="physics", description="Model type discriminator"
-    )
-    gen: Optional[GEN_TYPE] = Field(default=None)
-    sswell: Optional[SSWELL_TYPE] = Field(default=None)
-    negatinp: Optional[NEGATINP_TYPE] = Field(default=None)
-    wcapping: Optional[WCAPPING_TYPE] = Field(default=None)
-    quadrupl: Optional[QUADRUPL_TYPE] = Field(default=None)
-    breaking: Optional[BREAKING_TYPE] = Field(default=None)
-    friction: Optional[FRICTION_TYPE] = Field(default=None)
-    triad: Optional[TRIAD_TYPE] = Field(default=None)
-    vegetation: Optional[VEGETATION_TYPE] = Field(default=None)
-    mud: Optional[MUD_TYPE] = Field(default=None)
-    sice: Optional[SICE_TYPE] = Field(default=None)
-    turbulence: Optional[TURBULENCE_TYPE] = Field(default=None)
-    bragg: Optional[BRAGG_TYPE] = Field(default=None)
-    limiter: Optional[LIMITER_TYPE] = Field(default=None)
-    obstacle: Optional[OBSTACLE_TYPE] = Field(default=None)
-    setup: Optional[SETUP_TYPE] = Field(default=None)
-    diffraction: Optional[DIFFRACTION_TYPE] = Field(default=None)
-    surfbeat: Optional[SURFBEAT_TYPE] = Field(default=None)
-    scat: Optional[SCAT_TYPE] = Field(default=None)
-    deactivate: Optional[DEACTIVATE_TYPE] = Field(default=None)
-
-    # @validator("off")
-    # def deactivate_physics(cls, value, values):
-    #     if "windgrowth" in value:
-    #         logger.info("Switching off WINDGROWTH")
-    #     if "quadrupl" in value and values.get("quadrupl") is not None:
-    #         logger.warning("Switching off QUADRUPL")
-    #     if "wcapping" in value:
-    #         logger.info("Switching off WCAPPING")
-    #     if "breaking" in value:
-    #         logger.info("Switching off BREAKING")
-    #     return values
-
-    @model_validator(mode="after")
-    def negatinp_only_with_zieger(self) -> "PHYSICS":
-        """Log a warning if NEGATINP is used with a non-ZIEGER SSWELL."""
-        if self.negatinp is None:
-            return self
-        elif self.sswell is None:
-            logger.warning(
-                "The negative wind input NEGATINP is only intended to use with the "
-                "swell dissipation SSWELL ZIEGER but no SSWELL has been specified."
-            )
-        elif self.sswell.model_type != "zieger":
-            logger.warning(
-                "The negative wind input NEGATINP is only intended to use with the "
-                "swell dissipation SSWELL ZIEGER but the SSWELL "
-                f"{self.sswell.model_type.upper()} has been specified."
-            )
-        return self
-
-    def cmd(self):
-        repr = [self.gen.render()]
-        if self.sswell is not None:
-            repr += [f"{self.sswell.render()}"]
-        if self.negatinp is not None:
-            repr += [self.negatinp.render()]
-        if self.wcapping is not None:
-            repr += [self.wcapping.render()]
-        if self.quadrupl is not None:
-            repr += [self.quadrupl.render()]
-        if self.breaking is not None:
-            repr += [self.breaking.render()]
-        if self.friction is not None:
-            repr += [self.friction.render()]
-        if self.triad is not None:
-            repr += [self.triad.render()]
-        if self.vegetation is not None:
-            repr += [self.vegetation.render()]
-        if self.mud is not None:
-            repr += [self.mud.render()]
-        if self.sice is not None:
-            repr += [self.sice.render()]
-        if self.turbulence is not None:
-            repr += [self.turbulence.render()]
-        if self.bragg is not None:
-            repr += [self.bragg.render()]
-        if self.limiter is not None:
-            repr += [self.limiter.render()]
-        if self.obstacle is not None:
-            repr += self.obstacle.render()  # Object returns a list of components
-        if self.setup is not None:
-            repr += [self.setup.render()]
-        if self.diffraction is not None:
-            repr += [self.diffraction.render()]
-        if self.surfbeat is not None:
-            repr += [self.surfbeat.render()]
-        if self.scat is not None:
-            repr += [self.scat.render()]
-        if self.deactivate is not None:
-            repr += self.deactivate.render()  # Object returns a list of components
-        return repr
