+++ tests/swan/swan_model.yml
@@ -23,164 +23,164 @@ coordinates:
   kind:
     model_type: spherical
 
-cgrid:
-  model_type: regular
-  spectrum:
-    mdc: 36
-    flow: 0.04
-    fhigh: 0.4
-  xlenc: 100.0
-  ylenc: 100.0
-  mxc: 10
-  myc: 10
+# cgrid:
+#   model_type: regular
+#   spectrum:
+#     mdc: 36
+#     flow: 0.04
+#     fhigh: 0.4
+#   xlenc: 100.0
+#   ylenc: 100.0
+#   mxc: 10
+#   myc: 10
 
-inpgrid:
-  inpgrids:
-  - model_type: regular
-    grid_type: bottom
-    xpinp: 0.0
-    ypinp: 0.0
-    alpinp: 0.0
-    mxinp: 10
-    myinp: 10
-    dxinp: 0.1
-    dyinp: 0.1
-    excval: -999.0
-    readinp:
-      model_type: readinp
-      fname1: bottom.txt
-  - model_type: regular
-    grid_type: wind
-    xpinp: 0.0
-    ypinp: 0.0
-    alpinp: 0.0
-    mxinp: 10
-    myinp: 10
-    dxinp: 0.1
-    dyinp: 0.1
-    excval: -999.0
-    readinp:
-      model_type: readinp
-      fname1: wind.txt
-    nonstationary:
-      tbeg: '2023-01-01T00:00:00'
-      delt: PT30M
-      tend: '2023-02-01T00:00:00'
-      deltfmt: hr
+# inpgrid:
+#   inpgrids:
+#   - model_type: regular
+#     grid_type: bottom
+#     xpinp: 0.0
+#     ypinp: 0.0
+#     alpinp: 0.0
+#     mxinp: 10
+#     myinp: 10
+#     dxinp: 0.1
+#     dyinp: 0.1
+#     excval: -999.0
+#     readinp:
+#       model_type: readinp
+#       fname1: bottom.txt
+#   - model_type: regular
+#     grid_type: wind
+#     xpinp: 0.0
+#     ypinp: 0.0
+#     alpinp: 0.0
+#     mxinp: 10
+#     myinp: 10
+#     dxinp: 0.1
+#     dyinp: 0.1
+#     excval: -999.0
+#     readinp:
+#       model_type: readinp
+#       fname1: wind.txt
+#     nonstationary:
+#       tbeg: '2023-01-01T00:00:00'
+#       delt: PT30M
+#       tend: '2023-02-01T00:00:00'
+#       deltfmt: hr
 
-boundary:
-  model_type: boundspec
-  shapespec:
-    model_type: shapespec
-    shape:
-      model_type: tma
-      d: 12.0
-  location:
-    model_type: side
-    side: west
-  data:
-    model_type: constantpar
-    hs: 1.0
-    per: 10.0
-    dir: 0.0
-    dd: 10.0
+# boundary:
+#   model_type: boundspec
+#   shapespec:
+#     model_type: shapespec
+#     shape:
+#       model_type: tma
+#       d: 12.0
+#   location:
+#     model_type: side
+#     side: west
+#   data:
+#     model_type: constantpar
+#     hs: 1.0
+#     per: 10.0
+#     dir: 0.0
+#     dd: 10.0
 
-initial:
-  model_type: initial
-  kind:
-    model_type: hotsingle
-    fname: hotfile.txt
-    format: free
+# initial:
+#   model_type: initial
+#   kind:
+#     model_type: hotsingle
+#     fname: hotfile.txt
+#     format: free
 
-physics:
-  gen:
-    model_type: gen3
-    source_terms:
-      model_type: st6c1
-  sswell:
-    model_type: zieger
-  negatinp:
-    model_type: negatinp
-    rdcoef: 0.1
-  wcapping:
-    model_type: ab
-    cds2: 5.0e-5
-    br: 1.75e-3
-    current: True
-    cds3: 0.8
-  quadrupl:
-    iquad: 2
-    lambd: 0.25
-  breaking:
-    model_type: constant
-    alpha: 1.0
-    gamma: 0.73
-  friction:
-    model_type: collins
-    cfw: 0.015
-  triad:
-    model_type: dcta
-    trfac: 4.4
-  vegetation:
-    model_type: vegetation
-    height: 1.0
-    diamtr: 0.1
-    nstems: 1
-    drag: 0.1
-  mud:
-    model_type: mud
-    layer: 2.0
-    rhom: 1300.0
-    viscm: 0.0076
-  sice:
-    model_type: R19
-  turbulence:
-    ctb: 0.01
-    current: True
-    tbcur: 0.004
-  bragg:
-    model_type: ft
-    ibrag: 1
-    nreg: 200
-    cutoff: 5.0
-  limiter:
-    model_type: limiter
-    ursell: 10.0
-    qb: 1.0
-  obstacle:
-    model_type: obstacles
-    obstacles:
-      - model_type: obstacle
-        transmission:
-          model_type: transm
-          trcoef: 0.5
-        reflection:
-          reflc: 0.5
-        line:
-          xp: [174.1, 174.2, 174.3]
-          yp: [-39.1, -39.1, -39.1]
-      - model_type: obstacle
-        reflection:
-          reflc: 1.0
-        line:
-          xp: [174.3, 174.3]
-          yp: [-39.1, -39.3]
-      - model_type: fig
-        alpha1: 5e-4
-        hss: 2.1
-        tss: 12.0
-        line:
-          xp: [174.1, 174.2, 174.3]
-          yp: [-39.1, -39.1, -39.1]
-  setup:
-    supcor: 0.0
-  diffraction:
-    idiffr: 1
-  surfbeat:
-    spacing: uniform
-  scat:
-    model_type: scat
-  deactivate:
-    physics:
-      - physics: wcapping
-      - physics: breaking
+# physics:
+#   gen:
+#     model_type: gen3
+#     source_terms:
+#       model_type: st6c1
+#   sswell:
+#     model_type: zieger
+#   negatinp:
+#     model_type: negatinp
+#     rdcoef: 0.1
+#   wcapping:
+#     model_type: ab
+#     cds2: 5.0e-5
+#     br: 1.75e-3
+#     current: True
+#     cds3: 0.8
+#   quadrupl:
+#     iquad: 2
+#     lambd: 0.25
+#   breaking:
+#     model_type: constant
+#     alpha: 1.0
+#     gamma: 0.73
+#   friction:
+#     model_type: collins
+#     cfw: 0.015
+#   triad:
+#     model_type: dcta
+#     trfac: 4.4
+#   vegetation:
+#     model_type: vegetation
+#     height: 1.0
+#     diamtr: 0.1
+#     nstems: 1
+#     drag: 0.1
+#   mud:
+#     model_type: mud
+#     layer: 2.0
+#     rhom: 1300.0
+#     viscm: 0.0076
+#   sice:
+#     model_type: R19
+#   turbulence:
+#     ctb: 0.01
+#     current: True
+#     tbcur: 0.004
+#   bragg:
+#     model_type: ft
+#     ibrag: 1
+#     nreg: 200
+#     cutoff: 5.0
+#   limiter:
+#     model_type: limiter
+#     ursell: 10.0
+#     qb: 1.0
+#   obstacle:
+#     model_type: obstacles
+#     obstacles:
+#       - model_type: obstacle
+#         transmission:
+#           model_type: transm
+#           trcoef: 0.5
+#         reflection:
+#           reflc: 0.5
+#         line:
+#           xp: [174.1, 174.2, 174.3]
+#           yp: [-39.1, -39.1, -39.1]
+#       - model_type: obstacle
+#         reflection:
+#           reflc: 1.0
+#         line:
+#           xp: [174.3, 174.3]
+#           yp: [-39.1, -39.3]
+#       - model_type: fig
+#         alpha1: 5e-4
+#         hss: 2.1
+#         tss: 12.0
+#         line:
+#           xp: [174.1, 174.2, 174.3]
+#           yp: [-39.1, -39.1, -39.1]
+#   setup:
+#     supcor: 0.0
+#   diffraction:
+#     idiffr: 1
+#   surfbeat:
+#     spacing: uniform
+#   scat:
+#     model_type: scat
+#   deactivate:
+#     physics:
+#       - physics: wcapping
+#       - physics: breaking
